requirements.txt
black==22.3.0
discord.py==2.0.1
python-dotenv==0.21.0
